# home-banking/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Swagger admin
  description: >-
    This is a sample server admin server.  You can find out more about    
    Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net,
    #swagger](http://swagger.io/irc/).      For this sample, you can use the api
    key `special-key` to test the authorization     filters.
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://admin.swagger.io/v2'
  - url: 'http://admin.swagger.io/v2'
...
paths:
...
  "/homeBanking/account/signup":
    post:
      tags:
        - Account
      summary: account da registrare
      operationId: registra account
      requestBody:
        description: >-
          dati dell account da registrare, che fa riferimento ai dati di un
          cliente precedentemente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
        required: true
      responses:
        '200':
          description: registrazione account effettuata
        '400':
          description: dati non validi
          content: {}
        '403':
          description: account già esistente
  "/homeBanking/account/login/userpass":
    post:
      tags:
        - Account
      summary: Accedi all'account tramite username e password
      operationId: loginAccountUP
      requestBody:
        description: 'Credenziali per l''accesso: username e password'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credenziali'
        required: true
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
          content: {}
        '404':
          description: user not found
          content: {}
  "/homeBanking/account/login/fingerprint":
    post:
      tags:
        - "Account"
      summary: Accedi all'account tramite username e password
      operationId: loginAccountF
      requestBody:
        description: Accesso tramite impronta digitale
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Impronta'
        required: true
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
          content: {}
        '404':
          description: user not found
          content: {}
  "/homeBanking/account/cartaprepagata":
    post:
      tags:
        - "Carta Prepagata"
      summary: Aggiungi una nuova carta prepagata
      operationId: newCartaPre
      requestBody:
        description: Associa una nuova carta prepagata all'account
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartaPrepagata'
        required: true
      responses:
        '200':
          description: Carta associata con successo
        '400':
          description: Errore
  "/homeBanking/account/cartaprepagata/one":   
    post:
      tags:
        - "Carta Prepagata"
      summary: "Visualizza la carta prepagata"
      operationId: "srchOneCartaPre"
      requestBody:
        description: "Restituisce la carta prepagata a partire dal numero fornito"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NumeroCarta'
        required: true
      responses:
        '200':
          description: "Carta trovato con successo"
        '400':
          description: "Errore"
    get:
      tags: 
      - "Carta Prepagata"
      summary: "Visualizza tutte le carte prepagate"
      description: "Restituisce tutte le carte prepagate associate all'account"
      operationId: "srchAllCartePre"
      responses:
        '200':
          description: "Operazione corretta"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartaPrepagata'
    put:
      tags: 
      - "Carta Prepagata"
      summary: "Effettua un bonifico"
      description: "Effettua un bonifico su una carta"
      operationId: "doBonif"
      requestBody:
        description: "I dati per il bonifico: la carta del proprietario e del destinatario e l'importo"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatiBonificoCP'
        required: true
      responses:
        '200':
          description: "Bonifico effettuato con successo!"
...
components:
  schemas:
    Cliente:
      type: object
      properties:
        codice_fiscale:
          type: string
        nome:
          type: string
        cognome:
          type: string
        citta_di_nascita:
          type: string
        data_di_nascita:
          type: string
          format: date
        numero_di_telefono:
          type: string
        indirizzo_di_residenza:
          type: string
        citta_di_residenza:
          type: string
    Account:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        password:
          type: string
        email:
          type: string
        imponta_digitale:
          type: integer
        dispositivi_associati:
          type: string
        cliente:
          $ref: '#/components/schemas/Cliente'
    CartaPrepagata:
      type: object
      properties:
        numero:
          type: number
          format: double
        saldo_contabile:
          type: number
          format: double
        saldo_disponibile:
          type: number
          format: double
        data_di_scadenza:
          type: string
          format: date
        cvv:
          type: integer
        pin:
          type: integer
        account:
          $ref: '#/components/schemas/Account'
...
    Impronta:
      type: object
      properties:
        impronta:
          type: number
          format: long
    Credenziali:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    NumeroCarta:
      type: object
      properties:
        numero:
          type: string
    DatiBonificoCP:
      properties:
        carta_proprietario:
          $ref: '#/components/schemas/CartaPrepagata'
        carta_destinatario:
          $ref: '#/components/schemas/CartaPrepagata'
        importo:
          type: number
          format: double
...